refactor(tabs): split Tabs.module.scss into separate stylesheets
