+++ packages/tabs/src/Tab.module.scss
@@ -3,15 +3,6 @@
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/typography";
 
-.tabList {
-  border-bottom: 1px solid rgba($color-gray-600-rgb, 0.1);
-  padding: $spacing-xs $spacing-md 0;
-}
-
-.tabList.noPadding {
-  padding: 0;
-}
-
 .tab {
   display: inline-flex;
   align-items: center;
@@ -82,17 +73,6 @@
   align-items: center;
 }
 
-.tabPanel {
-  border: 2px solid transparent;
-
-  &:focus {
-    outline: none;
-  }
-  &:global(.focus-visible) {
-    border-color: $color-blue-500;
-    border-radius: $border-focus-ring-border-radius;
-  }
-}
 
 @media (forced-colors: active) {
   .tab {
@@ -124,7 +104,4 @@
     border-bottom: 2px solid transparent;
   }
 
-  .tabPanel {
-    border: 0;
-  }
 }
