+++ packages/tabs/src/TabList.module.scss
@@ -1,0 +1,11 @@
+@import "~@kaizen/design-tokens/sass/spacing";
+@import "~@kaizen/design-tokens/sass/color";
+
+.tabList {
+  border-bottom: 1px solid rgba($color-gray-600-rgb, 0.1);
+  padding: $spacing-xs $spacing-md 0;
+}
+
+.tabList.noPadding {
+  padding: 0;
+}
