+++ packages/tabs/src/TabList.tsx
@@ -1,7 +1,8 @@
 import { TabList as ReachTabList } from "@reach/tabs"
 import React, { ReactNode } from "react"
 import classnames from "classnames"
-import styles from "./Tabs.module.scss"
+// import styles from "./TabList.module.scss"
+import styles from "./TabList.module.scss"
 
 export interface TabListProps {
   /**
