+++ packages/tabs/src/TabPanel.module.scss
@@ -1,0 +1,14 @@
+@import "~@kaizen/design-tokens/sass/border";
+@import "~@kaizen/design-tokens/sass/color";
+
+.tabPanel {
+  border: 2px solid transparent;
+
+  &:focus {
+    outline: none;
+  }
+  &:global(.focus-visible) {
+    border-color: $color-blue-500;
+    border-radius: $border-focus-ring-border-radius;
+  }
+}
